9818: Rename integration test, start on testing electronic flow for econsent
